+++ Content.Shared/Vehicle/Components/GenericKeyedVehicleComponent.cs
@@ -1,4 +1,5 @@
-﻿using Content.Shared.Whitelist;
+﻿using Content.Shared.Vehicle.Systems;
+using Content.Shared.Whitelist;
 using Robust.Shared.GameStates;
 
 namespace Content.Shared.Vehicle.Components;
