+++ Content.Shared/Vehicle/Components/StrapVehicleComponent.cs
@@ -1,4 +1,5 @@
-﻿using Robust.Shared.GameStates;
+﻿using Content.Shared.Vehicle.Systems;
+using Robust.Shared.GameStates;
 
 namespace Content.Shared.Vehicle.Components;
