+++ Content.Shared/Vehicle/Systems/VehicleSystem.Key.cs
@@ -2,7 +2,7 @@ using System.Linq;
 using Content.Shared.Vehicle.Components;
 using Robust.Shared.Containers;
 
-namespace Content.Shared.Vehicle;
+namespace Content.Shared.Vehicle.Systems;
 
 public sealed partial class VehicleSystem
 {
